Emulate BiBTeX Buffer line breaking bug
When BiBTeX's output buffer breaks lines, it is supposed to start the next line
with two spaces. However the exact BiBTeX implementation contains a bug:
Whenever at least two whitespaces are found at the exact point where the
linebreak is supposed to occur, it adds an additional third space to the next
line. This commit emulates BiBTeX's exact behavior.